Supports operator overloading with return value optimization (obj + obj = create one object).

+++ test/fixtures/simple/include/types.h
@@ -4,6 +4,6 @@
 /** A simple typedef to have something not in any
  * class or namespace.
  */
-typedef float MyFloat;
+typedef double MyFloat;
 
 #endif // SIMPLE_INCLUDE_TYPES_H_
